Refactor of IdService to apply to multiple entities
